#10072: Merge tanhshrink_bw to TTNN
