+++ docs/source/ttnn/ttnn/dependencies/tt_lib.rst
@@ -860,8 +860,6 @@ Backward Operations
 
 .. autofunction:: tt_lib.tensor.reciprocal_bw
 
-.. autofunction:: tt_lib.tensor.tanhshrink_bw
-
 .. autofunction:: tt_lib.tensor.threshold_bw
 
 .. autofunction:: tt_lib.tensor.logiteps_bw
