+++ tt_eager/tt_dnn/op_library/backward/backward_ops.hpp
@@ -243,7 +243,7 @@ std::vector<Tensor> prod_bw(
     int64_t dim,
     const MemoryConfig& output_mem_config = operation::DEFAULT_OUTPUT_MEMORY_CONFIG);
 
-std::vector<Tensor> tanhshrink_bw(
+std::vector<Tensor> log_sigmoid_bw(
     const Tensor& grad,
     const Tensor& input,
     const MemoryConfig& output_mem_config = operation::DEFAULT_OUTPUT_MEMORY_CONFIG);
